+++ GXRuntime/graphics/frontend/include/gxruntime/aurora_recomp/retail_gx_frontend.hpp
@@ -138,7 +138,7 @@ private:
   TraceEventObserver event_observer_ = nullptr;
   void* event_observer_user_ = nullptr;
   std::uint32_t emitted_trace_count_ = 0;
-  std::uint32_t next_packet_sequence_ = 0;
+  std::uint64_t next_packet_sequence_ = 0;
   const char* last_error_ = nullptr;
   std::size_t last_error_offset_ = 0;
   std::uint8_t last_error_opcode_ = 0;
