+++ GXRuntime/graphics/frontend/src/retail_gx_frontend.cpp
@@ -864,7 +864,7 @@ bool RetailGxFrontend::emit_new_packets(AuroraRenderSink& sink,
     return false;
 
   for (std::uint32_t i = first_event; i < events.size(); ++i) {
-    const std::uint32_t sequence =
+    const std::uint64_t sequence =
         packet_drain_enabled_ ? next_packet_sequence_++ : i;
     RenderPacket packet = make_render_packet(sequence, events[i]);
     // Draw events are emitted exactly once, in order, so the retained payloads
